added always open as admin

+++ Powershell-Here.ps1
@@ -15,7 +15,8 @@ Create Powershell Here in right click menu
 Ensure you have admin rights or sufficient rights to modify registry
 
 .DATE CREATED
-30th August 2016
+30th Aug 2016 - opens powershell on right click.
+18TH Sep 2016 - opens powershell as admin on right click.
 
 .DEVELOPED BY
 Nitish Janawadkar
@@ -29,31 +30,32 @@ $ErrorActionPreference = "stop"
 
 try
 {
-    #GET THE PATH TO POWERSHELL EXECUTABLE AND CREATE THE COMMAND
-    $pspath = "$PSHome\powershell.exe -Noexit -Nologo"
-
     #COPY POWERSHELL ICON TO THE POWERSHELL EXECUTABLE LOCATION
-    Copy-Item .\Icons\powershell_5.ico -Destination $PSHome -Force
-    Write-Host "[OK] " -NoNewline -ForegroundColor Green
-    Write-Host "ADDED POWERSHELL ICON FOR THE MENU"
+    if(-not (Test-Path "$PSHome\powershell_5.ico")) { 
+        Copy-Item .\Icons\powershell_5.ico -Destination $PSHome -Force 
+        Write-Host "[OK] " -NoNewline -ForegroundColor Green
+        Write-Host "ADDED POWERSHELL ICON FOR THE MENU"
+    }
+    else
+    {
+        Write-Host "[OK] " -NoNewline -ForegroundColor Green
+        Write-Host "POWERSHELL ICON FOR THE MENU ALREADY EXISTS"
+    }
 
-    #ADD POWERSHELL HERE WHEN RIGHT CLICK IS DONE ON A FOLDER
-    New-Item HKLM:\SOFTWARE\Classes\Directory\shell\PSOpenHere -force | Out-Null
-    Set-Item HKLM:\SOFTWARE\Classes\Directory\shell\PSOpenHere "PowerShell Here"
-    New-ItemProperty -Path  HKLM:\SOFTWARE\Classes\Directory\shell\PSOpenHere -Name "Icon" -PropertyType "String" -Value "$PSHome\powershell_5.ico" -Force | Out-Null
-    New-item HKLM:\SOFTWARE\Classes\Directory\shell\PSOpenHere\command -force | Out-Null
-    Set-item HKLM:\SOFTWARE\Classes\Directory\shell\PSOpenHere\command "$pspath -Command Set-Location '%L'"
-    Write-Host "[OK] " -NoNewline -ForegroundColor Green
-    Write-Host "ADDED POWERSHELL HERE TO THE MENU WHEN RIGHT CLICK IS DONE ON ANY FOLDER"
+    #GET THE PATH TO POWERSHELL EXECUTABLE AND CREATE THE COMMAND
+    $menu = 'PowerShell Here (Admin)'
+    $command = "$PSHOME\powershell.exe -NoExit -NoProfile -Command ""Set-Location '%V'"""
+ 
+    'directory', 'directory\background', 'drive' | ForEach-Object {
+        New-Item -Path "Registry::HKEY_CLASSES_ROOT\$_\shell" -Name runas\command -Force |
+        Set-ItemProperty -Name '(default)' -Value $command -PassThru |
+        Set-ItemProperty -Path {$_.PSParentPath} -Name '(default)' -Value $menu -PassThru |
+        Set-ItemProperty -Name HasLUAShield -Value ''
+        New-ItemProperty -Path  "Registry::HKEY_CLASSES_ROOT\$_\shell\runas" -Name "Icon" -PropertyType "String" -Value "$PSHome\powershell_5.ico" -Force | Out-Null
 
-    #ADD POWERSHELL HERE WHEN RIGHT CLICK IS DONE IN THE EXPLORER WINDOW
-    New-Item HKLM:\SOFTWARE\Classes\Directory\Background\shell\PSOpenHere -force | Out-Null
-    Set-Item HKLM:\SOFTWARE\Classes\Directory\Background\shell\PSOpenHere "PowerShell Here"
-    New-ItemProperty -Path  HKLM:\SOFTWARE\Classes\Directory\Background\shell\PSOpenHere -Name "Icon" -PropertyType "String" -Value "$PSHome\powershell_5.ico" -Force | Out-Null
-    New-item HKLM:\SOFTWARE\Classes\Directory\Background\shell\PSOpenHere\command -force | Out-Null
-    Set-item HKLM:\SOFTWARE\Classes\Directory\Background\shell\PSOpenHere\command "$pspath -Command Set-Location '%v'"
-    Write-Host "[OK] " -NoNewline -ForegroundColor Green
-    Write-Host "ADDED POWERSHELL HERE TO THE MENU WHEN RIGHT CLICK IS DONE ON ANY EXPLORER WINDOW`n"
+        Write-Host "[OK] " -NoNewline -ForegroundColor Green
+        Write-Host "ADDED POWERSHELL HERE (ADMIN) TO THE MENU WHEN RIGHT CLICK IS DONE ON ANY" $_.ToUpper()
+    }
 }
 catch
 {
